requirements.txt
absl-py==0.13.0
affine==2.3.0
ansicon==1.89.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.10.0
blessed==1.18.1
blis==0.7.4
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cymem==2.0.5
daphne==3.0.2
dj-database-url==0.5.0
Django==3.0.14
django-crispy-forms==1.13.0
django-filter==21.1
django-jinja @ https://github.com/niwinz/django-jinja/archive/7d08e5d5d7cad5c00b65fd7c4b4f095f209a4162.tar.gz
django-rest-framework==0.1.0
djangorestframework==3.12.4
dominate==2.6.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.40.0
gunicorn==20.1.0
h5py==3.1.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.5.0
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
jinxed==1.1.0
joblib==1.1.0
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pathy==0.6.0
pbr==5.6.0
pefile==2021.9.3
prefixed==0.3.2
preshed==3.0.5
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyinstaller-hooks-contrib==2021.3
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scapy==2.4.5
scikit-learn==1.0.1
scipy==1.7.1
service-identity==21.1.0
six==1.15.0
sklearn==0.0
smart-open==5.2.1
snuggs==1.4.7
soupsieve==2.2.1
spacy==3.1.2
spacy-legacy==3.0.8
SQLAlchemy==1.4.23
sqlparse==0.4.1
srsly==2.4.1
striprtf==0.0.13
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thinc==8.0.10
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.2
treelib==1.6.1
txaio==21.2.1
typer==0.3.2
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.26.6
visitor==0.1.3
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
wwo-hist==0.0.7

